requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.5
crytic-compile==0.2.3
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
graphviz==0.20
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
prettytable==3.4.1
protobuf==3.20.2
pycryptodome==3.15.0
pyrsistent==0.18.1
pysha3==1.0.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
slither-analyzer==0.8.3
solc-select==0.2.1
requests==2.27.1
web3==5.30.0
colorama==0.4.5
mythril==0.23.16
